cleaned up TournamentPage for readability, tried to fix linting errors
